Create ProposalSkipCallbackHandler and update Telegram Bot

+++ src/main/java/org/l4ncet/telegrambot/dto/CreateOrderRequestDto.java
@@ -8,6 +8,6 @@ import java.time.LocalDate;
 public class CreateOrderRequestDto {
     String title;
     String description;
-    Long price;
+    Integer price;
     LocalDate deadline;
 }
